Added collaboration card
